libs/shared/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pymysql==1.1.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic[email]
pydantic-settings>=2.2
cryptography>=42,<43
